requirements.txt
click==6.7
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
itsdangerous==0.24
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.4.0
SQLAlchemy==1.3.6
Werkzeug==1.0.1
pytest==8.2.0